When importing visits on a day different from the visits day, invalidate the archived reports
